2to3

+++ edexOsgi/com.raytheon.edex.plugin.gfe/utility/common_static/base/gfe/python/AbsTime.py
@@ -1,19 +1,19 @@
 ##
 # This software was developed and / or modified by Raytheon Company,
-# pursuant to Contract DG133W-05-CQ-1067 with the US Government.
-# 
-# U.S. EXPORT CONTROLLED TECHNICAL DATA
+# pursuant to Contract DG133W-05-CQ-1067 with the US Government.
+# 
+# U.S. EXPORT CONTROLLED TECHNICAL DATA
 # This software product contains export-restricted data whose
 # export/transfer/disclosure is restricted by U.S. law. Dissemination
 # to non-U.S. persons whether in the United States or abroad requires
 # an export license or other authorization.
 # 
-# Contractor Name:        Raytheon Company
-# Contractor Address:     6825 Pine Street, Suite 340
-#                         Mail Stop B8
-#                         Omaha, NE 68106
-#                         402.291.0100
-# 
+# Contractor Name:        Raytheon Company
+# Contractor Address:     6825 Pine Street, Suite 340
+#                         Mail Stop B8
+#                         Omaha, NE 68106
+#                         402.291.0100
+# 
 # See the AWIPS II Master Rights File ("Master Rights File.pdf") for
 # further licensing information.
 ##
@@ -55,7 +55,7 @@ class AbsTime(datetime, JUtil.JavaWrapperClass):
        if arg is None:
            return None;
        
-       if isinstance(arg, int) or isinstance(arg, long) or isinstance(arg, float):            
+       if isinstance(arg, int) or isinstance(arg, int) or isinstance(arg, float):            
            tmp = datetime.utcfromtimestamp(arg)
        else:
            tmp = datetime.utcfromtimestamp(arg.getTime() / 1000)
@@ -71,7 +71,7 @@ class AbsTime(datetime, JUtil.JavaWrapperClass):
         return AbsTime(calendar.timegm(dt.utctimetuple()))
     
     def __add__(self, arg):        
-        if isinstance(arg, int) or isinstance(arg, long) or isinstance(arg, float):
+        if isinstance(arg, int) or isinstance(arg, int) or isinstance(arg, float):
             offset = arg         
         key = {"seconds": offset}
         dt = datetime.__add__(self, timedelta(**key))
@@ -80,7 +80,7 @@ class AbsTime(datetime, JUtil.JavaWrapperClass):
     def __sub__(self, arg):
         if isinstance(arg, AbsTime):
             return self.unixTime() - arg.unixTime()
-        elif isinstance(arg, int) or isinstance(arg, long) or isinstance(arg, float):
+        elif isinstance(arg, int) or isinstance(arg, int) or isinstance(arg, float):
             offset = arg 
         key = {"seconds": offset}
         dt = datetime.__sub__(self, timedelta(**key))
